// dist/tmp/app/shared/interfaces/commerce.interface.ts
export type PurchaseType =
  'SystemLicense' |
  'SystemLicenseNoDelivery' |
  'OfflineLicense' |
  'OfflineLicenseNoDelivery' |
  'PrepaySystemLicense' |
  'PrepayOfflineLicense' |
  'Trial' |
  'DeliveryOnly' |
  'ChannelNoDelivery' |
  'RevenueOnly' |
  'PrepayAccess';

export type PaymentOption =
  'Trial' |
  'DeliveryOnly' |
  'PurchaseOnCredit' |
  'CreditCard' |
  'Hold';

export type PaymentType =
  'PurchaseOnCredit' |
  'CreditCard';

export type QuoteStatus = 'ACTIVE' | 'PENDING' | 'ORDERED' | 'EXPIRED' | 'CANCELLED';

export type TranscodeStatus = 'Submitted' | 'Completed' | 'Failed' | 'UrlError' | 'Deleted';

export type OrderStatus = 'ORDER' | 'REFUND';

export type EditableQuoteFields = 'bulkOrderId' | 'discount' | 'purchaseType';

export const quotesWithoutPricing: PurchaseType[] = [
  'Trial',
  'DeliveryOnly'
];

export const quotesAllowedToHaveFeesOnly: PurchaseType[] = [
  'RevenueOnly',
  'PrepayAccess'
];

// Base interfaces

export interface CommonCommerce extends Common {
  projects?: Project[];
}

export interface Project {
  id: string;
  name: string;
  clientName: string;
  subTotal?: number;
  creditMemoForProjectId?: number;
  lineItems?: Array<AssetLineItem>;
  feeLineItems?: Array<FeeLineItem>;
  attributes?: Array<SelectedPriceAttribute>;
  [index: string]: any;
}

export interface LineItemNote {
  notes: string[];
}

export interface AssetLineItem {
  id?: string;
  asset?: Asset;
  notes?: Array<LineItemNote>;
  itemPrice?: number;
  multiplier?: number;
  transcodeStatus?: TranscodeStatus;
  downloadUrl?: string;
  asperaSpec?: string;
  tomSubmitTimestamp?: number;
  tomCompleteTimestamp?: number;
  tomTransactionId?: string;
  selectedTranscodeTarget?: string;
  transcodeTargets?: Array<string>;
  expirationDate?: number;
  attributes?: Array<SelectedPriceAttribute>;
  salesForceId?: string;
  price?: number;
  grossAssetPrice?: number;
  rightsManaged?: string;
  externalAgreementIds?: string[];
}

export interface FeeLineItem {
  id?: string;
  amount?: string;
  feeType?: string;
  notes?: string;
  salesForceId?: string;
}

export interface Asset {
  assetId?: number;
  assetName?: string;
  assetDuration?: number;
  metadata?: Metadatum[];
  rightsManaged?: string;
  supplierId?: number;
  supplierName?: string;
  thumbnailUrl?: string;
  timeStart?: number;
  timeEnd?: number;
  clipUrl?: string;
  uuid?: string;
  price?: number;
  masterDownloadUrl?: string;
}

export interface Metadatum {
  name: string;
  value: string;
}
// ...
export interface Cart extends CommonCommerce {
  discount?: number;
  itemCount?: number;
  projects?: Project[];
  stripePublicKey?: string;
  subTotal?: number;
  total?: number;
  userId?: number;
}

export interface Order extends CommonCommerce {
  paymentTerms?: string;
  poNumber?: string;
  discount?: number;
  bulkOrderId?: string;
  createdUserId: number;
  ownerUserId: number;
  ownerData?: OwnerData;
  orderAddress?: OrderAddress;
  orderStatus: OrderStatus;
  orderType: PurchaseType;
  paymentType: PaymentType;
  quoteId: number;
  taxAmount: number;
  licenseAgreementId: string;
  refundAmount: number;
  salesVertical: string;
  oldCommerceId: number;
  salesForceId: string;
  createdByIntegration: boolean;
  salesForceSyncedError: boolean;
  paymentBalance: number;
  paymentDueDate?: Date;
  creditMemoForOrderId?: number;
  subTotal?: number;
  total?: number;
  totalDiscountAmount?: number;
}
// ...
export interface AddAssetParameters {
  lineItem: AssetLineItem;
  markers?: SubclipMarkers;
  attributes?: {
    [index: string]: any;
  };
}
// ...
export interface PurchaseOptions {
  poNumber?: string;
  orderAddressId?: number;
  orderAddressType?: string;
  stripeToken?: string;
  stripeTokenType?: string;
}

export interface PaymentOptions {
  paymentOptions: Array<PaymentOption>;
  explanation: string;
  noCheckout: boolean;
}

export interface LicenseAgreementDocument {
  name: string;
  text: string;
}

export interface LicenseAgreement {
  projectType?: string;
  rights?: string;
  matchingAssets: Array<LicenseAsset | EnhancedAsset>;
  document: LicenseAgreementDocument;
}

export interface LicenseAgreements {
  items: Array<LicenseAgreement>;
}

export interface LicenseAsset {
  assetId: number;
  assetLineItemId: string;
  name: string;
  thumbnailUrl: string;
}

export interface CreditCardAuthorization {
  card: {
    brand: string;
    last4: string;
    exp_month: string;
    exp_year: string;
  };
  error?: {
    code: number;
    param: any;
  };
  id?: string;
  type?: string;
}
export interface OwnerData {
  accountName?: string;
  email?: string;
  firstName?: string;
  lastName?: string;
}
export interface OrderAddress {
  accountId: number;
  accountName?: string;
  billingInfo?: { address: Address; };
  email?: string;
  firstName?: string;
  lastName?: string;
  type: 'User' | 'Account';
}